Multiple Changes
Added more clock speeds
Updated all tests
To-Do: Fix low values of inductance to not need hacky equations with imperial values.

+++ Libraries/wait.h
@@ -15,6 +15,6 @@
 // Subroutines
 //-----------------------------------------------------------------------------
 
-void waitMicrosecond(uint32_t us);
+void waitMicrosecond(uint64_t us);
 
 #endif
